Implemented differential reading when correct checksum, checksum comparison bug fixed
Now using a freeze out to bypass calculations for the next 4 bytes
after a sucessful checksum calculation (frames always 5 bytes long).
Fixed bug in checksum comparison. It really seems that the Nonin
documentation is wrong, they are not doing a mod 255 of the sum of the
4 data bytes to make the checksum byte. Instead they are just keeping
the last 2 characters of the hex of the calculated sum